Add and use makeBindingsParams() utility function
Call makeBindingsParams() in graphs tests
Call makeBindingsParams() from rows.query()
