+++ lib/graphs.js
@@ -19,6 +19,7 @@ var requester = require('./requester.js');
 var mlutil    = require('./mlutil.js');
 var Operation = require('./operation.js');
 var qb        = require('./query-builder.js');
+var mlutil    = require('./mlutil.js');
 
 function uriErrorTransform(message) {
   /*jshint validthis:true */
@@ -965,61 +966,7 @@ function makeCommonSPARQLParams(
     if (sep === '?') { sep = '&'; }
   }
   if (bindings !== null && bindings !== void 0) {
-    var keys = Object.keys(bindings);
-    var max = keys.length;
-
-    var key = null;
-    var obj = null;
-    var type = null;
-    var lang = null;
-    var hasType = false;
-    var hasLang = false;
-    var name = null;
-    var value = null;
-
-    for (var i = 0; i < max; i++) {
-      key = keys[i];
-      obj = bindings[key];
-
-      value = obj.value;
-      if (value === void 0) {
-        name  = key;
-        value = obj;
-      } else {
-        type = obj.type;
-        lang = obj.lang;
-
-        hasType = (type !== null && type !== void 0);
-        if (hasType) {
-          if (typeof type !== 'string' && !(type instanceof String)) {
-            throw new Error('type must be string');
-          }
-          if (type.indexOf(':') > -1) {
-            throw new Error('type cannot contain colon - '+type);
-          }
-        }
-
-        hasLang = (lang !== null && lang !== void 0);
-        if (hasLang) {
-          if (typeof lang !== 'string' && !(lang instanceof String)) {
-            throw new Error('lang must be string');
-          }
-        }
-
-        if (hasType && hasLang && type !== 'string') {
-          throw new Error('cannot combine type with lang - '+type+' '+lang);
-        } else if (hasLang) {
-          name = key + '@' + lang;
-        } else if (hasType) {
-          name = key + ':' + type;
-        } else {
-          name = key;
-        }
-      }
-
-      endpoint += sep+encodeURIComponent('bind:'+name)+'='+encodeURIComponent(value);
-      if (i === 0 && sep === '?') { sep = '&'; }
-    }
+    endpoint += mlutil.makeBindingsParams(bindings, sep);
   }
 
   return (endpoint === '') ? null : endpoint;
