+++ lib/mlutil.js
@@ -392,6 +392,69 @@ function makeSliceClause(variant, args) {
   return sliceClause;
 }
 
+function makeBindingsParams(bindings, sep) {
+
+  var keys = Object.keys(bindings);
+  var max = keys.length;
+
+  var key = null;
+  var obj = null;
+  var type = null;
+  var lang = null;
+  var hasType = false;
+  var hasLang = false;
+  var name = null;
+  var value = null;
+  var paramStr = '';
+
+  for (var i = 0; i < max; i++) {
+    key = keys[i];
+    obj = bindings[key];
+
+    value = obj.value;
+    if (value === void 0) {
+      name  = key;
+      value = obj;
+    } else {
+      type = obj.type;
+      lang = obj.lang;
+
+      hasType = (type !== null && type !== void 0);
+      if (hasType) {
+        if (typeof type !== 'string' && !(type instanceof String)) {
+          throw new Error('type must be string');
+        }
+        if (type.indexOf(':') > -1) {
+          throw new Error('type cannot contain colon - '+type);
+        }
+      }
+
+      hasLang = (lang !== null && lang !== void 0);
+      if (hasLang) {
+        if (typeof lang !== 'string' && !(lang instanceof String)) {
+          throw new Error('lang must be string');
+        }
+      }
+
+      if (hasType && hasLang && type !== 'string') {
+        throw new Error('cannot combine type with lang - '+type+' '+lang);
+      } else if (hasLang) {
+        name = key + '@' + lang;
+      } else if (hasType) {
+        name = key + ':' + type;
+      } else {
+        name = key;
+      }
+    }
+
+    paramStr += sep+encodeURIComponent('bind:'+name)+'='+encodeURIComponent(value);
+    if (i === 0 && sep === '?') { sep = '&'; }
+  }
+
+  return paramStr;
+
+}
+
 module.exports = {
     addTxidHeaders:      addTxidHeaders,
     appendItem:          appendItem,
@@ -406,6 +469,7 @@ module.exports = {
     first:               first,
     getTxidParam:        getTxidParam,
     makeSliceClause:     makeSliceClause,
+    makeBindingsParams:  makeBindingsParams,
     marshal:             marshal,
     multipartBoundary:   multipartBoundary,
     parseJSON:           parseJSON,
