+++ lib/rows.js
@@ -20,6 +20,7 @@
 var requester  = require('./requester.js');
 var mlutil     = require('./mlutil.js');
 var Operation  = require('./operation.js');
+var mlutil     = require('./mlutil.js');
 
 /** @ignore */
 function Rows(client) {
@@ -109,65 +110,8 @@ Rows.prototype.query = function queryRows(builtPlan, options) {
 
   var endpoint = '/v1/rows';
 
-  // Bindings handling similar to graphs#makeCommonSPARQLParams
-  var bindings = null;
   if (options.bindings) {
-    var bindings = options.bindings;
-    var keys = Object.keys(options.bindings);
-    var max = keys.length;
-
-    var key = null;
-    var obj = null;
-    var type = null;
-    var lang = null;
-    var hasType = false;
-    var hasLang = false;
-    var name = null;
-    var value = null;
-
-    for (var i = 0; i < max; i++) {
-      key = keys[i];
-      obj = bindings[key];
-
-      value = obj.value;
-      if (value === void 0) {
-        name  = key;
-        value = obj;
-      } else {
-        type = obj.type;
-        lang = obj.lang;
-
-        hasType = (type !== null && type !== void 0);
-        if (hasType) {
-          if (typeof type !== 'string' && !(type instanceof String)) {
-            throw new Error('type must be string');
-          }
-          if (type.indexOf(':') > -1) {
-            throw new Error('type cannot contain colon - '+type);
-          }
-        }
-
-        hasLang = (lang !== null && lang !== void 0);
-        if (hasLang) {
-          if (typeof lang !== 'string' && !(lang instanceof String)) {
-            throw new Error('lang must be string');
-          }
-        }
-
-        if (hasType && hasLang && type !== 'string') {
-          throw new Error('cannot combine type with lang - '+type+' '+lang);
-        } else if (hasLang) {
-          name = key + '@' + lang;
-        } else if (hasType) {
-          name = key + ':' + type;
-        } else {
-          name = key;
-        }
-      }
-
-      endpoint += sep+encodeURIComponent('bind:'+name)+'='+encodeURIComponent(value);
-      if (i === 0 && sep === '?') { sep = '&'; }
-    }
+    endpoint += mlutil.makeBindingsParams(options.bindings, sep);
   }
 
   var connectionParams = this.client.connectionParams;
